pyproject.toml
[project]
name = "new-detective"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.20",
    "ctgan>=0.11.0",
    "faiss-cpu==1.9.0",
    "ipykernel>=6.30.1",
    "markovify>=0.9.4",
    "matplotlib>=3.10.5",
    "networkx>=3.4.2",
    "ollama>=0.5.3",
    "pandas==2.2.3",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.4",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.3",
    "sdv>=1.26.0",
    "sentence-transformers==3.2.0",
    "torch>=2.8.0",
    "torch-geometric>=2.6.1",
]
